refactored zipcode functionality to be it's own component, better separation of concerns
