+++ client/redux/actions/changeModalOpen.js
@@ -1,0 +1,9 @@
+const changeModalOpen = (ModalOpen) => {
+
+    return ({
+        type: 'CHANGE_MODALOPEN',
+        ModalOpen: ModalOpen
+    })
+}
+
+export default changeModalOpen;
