+++ client/redux/containers/searchContainer.js
@@ -3,6 +3,8 @@ import SearchBar from '../../components/searchBar.jsx';
 import changeSearch from '../actions/changeSearch.js';
 import handleSearch from '../actions/handleSearch.js';
 import changeProduct from '../actions/changeProduct.js';
+import changeModalOpen from '../actions/changeModalOpen.js';
+
 const mapStateToProps = (state) => {
     
     return ({
@@ -14,7 +16,8 @@ const mapDispatchToProps = (dispatch, ownProps) => {
     return ({
             handleSearchChange: (ownProps) => dispatch(changeSearch(ownProps)),
             handleProductSearch: (ownProps) => dispatch(handleSearch(ownProps)),
-            deleteCurrentProduct: () => dispatch(changeProduct(null))
+            deleteCurrentProduct: () => dispatch(changeProduct(null)),
+            handleModalChange: (ownProps) => dispatch(changeModalOpen(ownProps))
         })
 }
