+++ client/redux/reducers.js
@@ -2,35 +2,41 @@ import Redux from 'redux';
 
 const initialState = {
     search: '',
-    searchResults: [1,2,3,4],
+    searchResults: null,
     zip: null,
-    currentProduct: null
+    currentProduct: null,
+    modalOpen: true
 }
 
 const rootReducer = (state = initialState, action) => {
-    switch (action.type) {
-        case 'CHANGE_SEARCH':
-            return Object.assign({}, state, {
-                search: action.search
-              });
-        
-        case 'CHANGE_SEARCHRESULTS':
-              return Object.assign({}, state, {
-                searchResults: action.searchResults
-              });
+      switch (action.type) {
+            case 'CHANGE_SEARCH':
+                  return Object.assign({}, state, {
+                  search: action.search
+                  });
+            
+            case 'CHANGE_SEARCHRESULTS':
+                  return Object.assign({}, state, {
+                  searchResults: action.searchResults
+                  });
 
-        case 'CHANGE_PRODUCT':
-              return Object.assign({}, state, {
-                currentProduct: action.product
-              });
+            case 'CHANGE_PRODUCT':
+                  return Object.assign({}, state, {
+                  currentProduct: action.product
+                  });
 
-        case 'CHANGE_ZIP':
-              return Object.assign({}, state, {
-                zip: action.zip
-              });
+            case 'CHANGE_ZIP':
+                  return Object.assign({}, state, {
+                  zip: action.zip
+                  });
 
-        default:
-              return state;
+            case 'CHANGE_MODALOPEN':
+                  return Object.assign({}, state, {
+                  modalOpen: action.ModalOpen
+                  });
+
+            default:
+                  return state;
     }
 }
